rename StyledFollowerListButton, unset navbar content on unmount
